requirements.txt
Pillow>=12.2.0
pillow-heif>=1.3.0
opencv-contrib-python>=4.13.0
mediapipe>=0.10.33
onnxruntime>=1.24.4
